requirements.txt
spacy~=3.7.4
joblib~=1.4.2
torch~=2.3.0
numpy
tqdm
transformers
bs4
faiss-cpu
dvc
dvc-gdrive
sentencepiece
unstructured
langchain-community
langchain~=0.2.1
sentence-transformers
datasets
llama-cpp-python
python-telegram-bot
python-dotenv~=1.0.1
